% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.idie_exposed.R
\name{print.idie_exposed}
\alias{print.idie_exposed}
\alias{print.tmle_exposed}
\title{Print function for \code{idie_exposed}}
\usage{
\method{print}{idie_exposed}(x, ...)
}
\description{
Print function for \code{idie_exposed} a targeted minimum-loss based estimator for the Interventional
Disparity Indirect Effect (IDIE) among the exposed.
}
\author{
Amalie Lykkemark Moller \email{amalielykkemark@live.dk}
}
