% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tmle_exposed.R
\name{print.tmle_exposed}
\alias{print.tmle_exposed}
\title{Print function for \code{tmle_exposed}}
\usage{
\method{print}{tmle_exposed}(x, ...)
}
\description{
Print function for \code{tmle_exposed} a targeted minimum-loss based estimator for stochastic and deterministic interventions among the exposed.
}
\author{
Amalie Lykkemark Moller \email{amalielykkemark@live.dk}
}
