% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.idie_exposed.R
\name{summary.idie_exposed}
\alias{summary.idie_exposed}
\title{Summary function for \code{idie_exposed}}
\usage{
\method{summary}{idie_exposed}(x, ...)
}
\description{
Summary function for \code{idie_exposed} a targeted minimum-loss based estimator for the Interventional
Disparity Indirect Effect (IDIE) among the exposed.
}
\author{
Amalie Lykkemark Moller \email{amalielykkemark@live.dk}
}
