% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tmle_exposed.R
\name{summary.tmle_exposed}
\alias{summary.tmle_exposed}
\title{Summary function for \code{tmle_exposed}}
\usage{
\method{summary}{tmle_exposed}(x, ...)
}
\description{
Summary function for \code{tmle_exposed} a targeted minimum-loss based estimator for stochastic and deterministic interventions among the exposed.
}
\author{
Amalie Lykkemark Moller \email{amalielykkemark@live.dk}
}
